add the tracking page to the dashboard
